Move annotations from Op to Node
