+++ src/ngraph/op/op.hpp
@@ -18,9 +18,7 @@
 
 #include <string>
 
-#include "ngraph/autodiff/adjoints.hpp"
 #include "ngraph/node.hpp"
-#include "ngraph/op/util/op_annotations.hpp"
 
 namespace ngraph
 {
@@ -30,15 +28,6 @@ namespace ngraph
         class Op : public Node
         {
         public:
-            void set_op_annotations(std::shared_ptr<ngraph::op::util::OpAnnotations> op_annotations)
-            {
-                m_op_annotations = op_annotations;
-            }
-            std::shared_ptr<ngraph::op::util::OpAnnotations> get_op_annotations() const
-            {
-                return m_op_annotations;
-            }
-
             virtual bool is_op() const override { return true; }
         protected:
             Op()
@@ -48,9 +37,6 @@ namespace ngraph
             Op(const NodeVector& arguments);
             Op(const OutputVector& arguments);
             Op(const std::string& node_type, const NodeVector& arguments);
-
-        private:
-            std::shared_ptr<ngraph::op::util::OpAnnotations> m_op_annotations;
         };
     }
 }
